+++ R/zzz.R
@@ -4,7 +4,11 @@ matchms_similarity <- NULL
 matchms_filtering <- NULL
 spectrum_utils <- NULL
 
-#' @importFrom reticulate import use_virtualenv use_condaenv py_available py_install virtualenv_exists virtualenv_create virtualenv_remove conda_list conda_create
+#' @importFrom reticulate import use_virtualenv use_condaenv
+#'
+#' @importFrom reticulate py_install virtualenv_exists virtualenv_create
+#'
+#' @importFrom reticulate conda_list conda_create
 .onLoad <- function(libname, pkgname) {
     envname <- .spectripy_env()
     use_conda <- .spectripy_use_conda()
@@ -84,8 +88,7 @@ spectrum_utils <- NULL
                        envname = envname, method = "conda", pip = FALSE,
                        channel = c("bioconda", "conda-forge"), ...)
         } else {
-            py_install(c("spectrum_utils==0.3.2",
-                         "numpy==2.0.2"),
+            py_install(c("spectrum_utils==0.3.2", "numpy==2.0.2"),
                        envname = envname, method = "virtualenv",
                        channel = c("conda-forge"), ...)
         }